ochestrator_agent/requirements.txt
google-adk
fastmcp
qdrant-client
google-cloud-aiplatform
google-auth
python-dotenv
httpx
uvicorn
a2a==0.44
a2a-sdk==0.3.22